Registration works! (For some reason)

+++ src/main/java/com/example/Appointment/exception/UserNotFoundException.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(value= HttpStatus.NOT_FOUND)
 public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(Long id){
+    public UserNotFoundException(Integer id){
         super("User with id \"" + id + "\" not found");
     }
 }
