+++ src/main/java/com/example/Appointment/repository/RoleRepository.java
@@ -1,8 +1,9 @@
 package com.example.Appointment.repository;
 
+
 import com.example.Appointment.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     Role findByName(String name);
 }
